LayoutTests/platform/ios-simulator/media/video-controls-captions-trackmenu-hide-on-click-expected.txt

Test that we are able to trigger the list of captions, and then click elsewhere to hide it.


EVENT(canplaythrough)

*** Set the user language preference.
RUN(internals.setUserPreferredLanguages(['en']))
undefined FAIL
END OF TEST
